backend/pyproject.toml
[project]
name = "todo-app-ph2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "cohere",
    "fastapi[standard]>=0.128.0",
    "openai-agents>=0.6.4",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "sqlmodel",
]
